% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{plsgenomics-deprecated}
\alias{plsgenomics-deprecated}
\alias{rirls.spls}
\alias{rirls.spls.tune}
\alias{spls.adapt}
\alias{spls.adapt.tune}
\alias{plsgenomics-deprecated-package}
\title{Deprecated function(s) in the 'plsgenomics' package}
\usage{
rirls.spls(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
These functions are provided for compatibility with older version of
the 'plsgenomics' package.  They may eventually be completely
removed.
}
\section{Details}{

\tabular{rl}{
    \code{rirls.spls} \tab is replaced by \code{\link{logit.spls}}\cr
    \code{rirls.spls.tune} \tab is replaced by \code{\link{logit.spls.cv}}\cr
    \code{spls.adapt} \tab is replaced by \code{\link{spls}}\cr
    \code{spls.adapt.tune} \tab is replaced by \code{\link{spls.cv}}\cr
}
}

