\name{plsgenomics-internal}
\alias{standard.simpls}
\alias{unitr.simpls}
\alias{transformy}
\alias{pls.regression.sample}
\alias{pls.lda.sample}

\title{Internal plsgenomics Functions}
\description{
  Internal plsgenomics functions.
}
\note{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{internal}
