% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{package}
\name{plsgenomics-internal}
\alias{plsgenomics-internal}
\alias{standard.simpls}
\alias{unitr.simpls}
\alias{transformy}
\alias{pls.regression.sample}
\alias{pls.lda.sample}
\alias{gsim.aux}
\alias{mgsimaux}
\alias{rplsaux}
\alias{mrplsaux}
\alias{hplugin}
\alias{wirrls}
\alias{mwirrls}
\alias{ust}
\alias{ust.adapt}
\alias{wpls}
\alias{spls.aux}
\alias{spls.in}
\alias{logit.spls.aux}
\alias{multinom.spls.aux}
\alias{safeExp}
\alias{safeExpMat}
\alias{safeSum}
\alias{softMax}
\alias{plsgenomics-internal-package}
\title{Internal Functions for the 'plsgenomics' package}
\description{
These are not to be called by the user.
}
\keyword{internal}
